# openapi.yml
openapi: 3.0.3
info:
  title: ClubManager API
  description: >
    Specification for the ClubManager API,
    used by web and mobile clients to interact with the server.
  version: 0.1.0
...
components:
  securitySchemes:
  # I'm not sure if we are going to need BasicAuth depends how we handle Cognito
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: [] 
paths:
...
  /api/member/new:
    post:
      tags:
        - member
      summary: Create a new member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/member/{memberID}:
    parameters:
      - name: memberID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - member
      summary: Get this member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - member
      summary: Update this member
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - member
      summary: Delete this member
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/member/list:
    get:
      tags:
        - member
      summary: Get all members
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: role
          in: query
          description: "Optional: filter members by a specific role"
          schema:
            type: string
            enum:
              - boardMember
              - member
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO

  # membership
...
  /api/job/new:
    post:
      tags:
        - job
      summary: Create a new job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/job/{jobID}:
    parameters:
      - name: jobID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - job
      summary: Get this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - job
      summary: Update this job
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.

        **Note 1:** This endpoint is used to verify work points.
        
        **Note 2:** While most fields require admin permissions
        to update, both *1)* assigning the job to a member or paid laborer and
        *2)* checking in to the job can be done by a member as well.
        
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    post:
      tags:
        - job
      summary: Clone this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - job
      summary: Delete this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/job/list:
    get:
      tags:
        - job
      summary: Get all jobs
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: assignmentStatus
          in: query
          description: "Optional: filter jobs by a specific assignment status"
          schema:
            type: string
            enum:
              - open
              - assigned
        - name: verificationStatus
          in: query
          description: "Optional: filter jobs by a specific verification status"
          schema:
            type: string
            enum:
              - pending
              - verified
        - name: memberID
          in: query
          description: "Optional: pull jobs for a specific **member**"
          schema:
            type: integer
        - name: membershipID
          in: query
          description: "Optional: pull jobs for a specific **membership**"
          schema:
            type: integer
        - name: eventID
          in: query
          description: "Optional: pull jobs for a specific **event**"
          schema:
            type: integer
        - name: Range
          in: header
          description: |
            Optional: filter jobs by a specific timeframe.
            Format is one of:
            - `YYYYMMDD-YYYYMMDD`
            - `YYYYMMDD-`
            - `-YYYYMMDD`
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        206:
          description: Partial Content
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/jobType/new:
    post:
      tags:
        - jobType
      summary: Create a new job type
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/jobType/{jobTypeID}:
    parameters:
      - name: jobTypeID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - jobType
      summary: Get this job type
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - jobType
      summary: Update this job type
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.

        **Note:** This endpoint is used to mark a job type as inactive, rather
        than using a `DELETE` endpoint to remove it entirely.
        
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/jobType/list:
    get:
      tags:
        - jobType
      summary: Get all job types
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  
  # workPoints
...